Move log writer to router
