+++ airframe-http/.jvm/src/main/scala/wvlet/airframe/http/HttpResponseCodec.scala
@@ -11,10 +11,9 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package wvlet.airframe.http.router
+package wvlet.airframe.http
 
 import wvlet.airframe.codec.{JSONCodec, MessageCodec, MessageContext}
-import wvlet.airframe.http.{HttpResponse, HttpResponseAdapter}
 import wvlet.airframe.msgpack.spi.{Packer, Unpacker}
 
 /**
